Move tests to the com.jetbrains.rider.plugins.fsharp.test.cases package
This will follow the structure we have in the dotnet-products repo, so
it is unified and clear among the projects.
