+++ rider-fsharp/src/test/kotlin/com/jetbrains/rider/plugins/fsharp/test/cases/editor/CodeFoldingTest.kt
