+++ rider-fsharp/src/test/kotlin/com/jetbrains/rider/plugins/fsharp/test/cases/typeProviders/TypeProvidersFeaturesTest.kt
@@ -1,0 +1,87 @@
+package com.jetbrains.rider.plugins.fsharp.test.cases.typeProviders
+
+import com.intellij.openapi.actionSystem.ActionPlaces
+import com.intellij.openapi.actionSystem.IdeActions
+import com.jetbrains.rdclient.testFramework.executeWithGold
+import com.jetbrains.rdclient.testFramework.waitForDaemon
+import com.jetbrains.rider.test.annotations.TestEnvironment
+import com.jetbrains.rider.test.base.EditorTestBase
+import com.jetbrains.rider.test.enums.CoreVersion
+import com.jetbrains.rider.test.enums.ToolsetVersion
+import com.jetbrains.rider.test.scriptingApi.*
+import org.testng.annotations.Test
+import java.time.Duration
+
+@Test
+@TestEnvironment(
+    toolset = ToolsetVersion.TOOLSET_17_CORE,
+    coreVersion = CoreVersion.DOT_NET_6
+)
+class TypeProvidersFeaturesTest : EditorTestBase() {
+    override fun getSolutionDirectoryName() = "SwaggerProviderCSharp"
+    override val restoreNuGetPackages = true
+
+    @Test
+    fun `signature file navigation`() = doNavigationTestWithMultipleDeclarations()
+
+    @Test
+    fun `provided member navigation`() = doNavigationTest("SwaggerProvider.fs")
+
+    @Test
+    fun `provided abbreviation navigation`() = doNavigationTest("SwaggerProvider.fs")
+
+    @Test
+    fun `navigate to decompiled`() {
+        withOpenedEditor("CSharpLibrary/CSharpLibrary.cs", "CSharpLibrary.cs") {
+            waitForDaemon()
+            callAction(IdeActions.ACTION_GOTO_SUPER)
+            waitForEditorSwitch("ProvidedApiClientBase.cs")
+        }
+    }
+
+    @Test
+    fun `provided nested type navigation`() = doNavigationTest("SwaggerProvider.fs")
+
+    @Test
+    fun `multiply different abbreviation type parts - before`() = doNavigationTestWithMultipleDeclarations()
+
+    @Test
+    fun `multiply different abbreviation type parts - after`() = doNavigationTestWithMultipleDeclarations()
+
+    @Test
+    fun `provided member rename disabled`() = doRenameUnavailableTest()
+
+    @Test
+    fun `provided nested type rename disabled`() = doRenameUnavailableTest()
+
+    private fun doNavigationTest(declarationFileName: String) {
+        withOpenedEditor("CSharpLibrary/CSharpLibrary.cs", "CSharpLibrary.cs") {
+            waitForDaemon()
+            gotoDeclaration {
+                waitForEditorSwitch(declarationFileName)
+                waitForDaemon()
+                executeWithGold(testGoldFile) {
+                    dumpOpenedDocument(it, project!!, true)
+                }
+            }
+        }
+    }
+
+    private fun doNavigationTestWithMultipleDeclarations() {
+        withOpenedEditor("CSharpLibrary/CSharpLibrary.cs", "CSharpLibrary.cs") {
+            waitForDaemon()
+            executeWithGold(testGoldFile) {
+                callActionAndHandlePopup(IdeActions.ACTION_GOTO_DECLARATION, it, true, Duration.ofSeconds(1)) {
+                    this.closeAll()
+                }
+            }
+        }
+    }
+
+    private fun doRenameUnavailableTest() {
+        withOpenedEditor("CSharpLibrary/CSharpLibrary.cs", "CSharpLibrary.cs") {
+            waitForDaemon()
+            assertActionDisabled(project!!, dataContext, ActionPlaces.UNKNOWN, IdeActions.ACTION_RENAME)
+        }
+    }
+}
